% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkHeight.R
\name{checkHeight}
\alias{checkHeight}
\title{Return the height difference of the input dataframe withe the mean height of the dataframe}
\usage{
checkHeight(students.input, sex.specific = TRUE, print.statement = FALSE)
}
\arguments{
\item{students.input}{a dataframe with columns: names, heigth, sex}

\item{sex.specific}{a boolean which indicates if the height difference is calculated sex specific}

\item{print.statement}{a boolean; TRUE prints a message after calculation is done}
}
\value{
A dataframe with column names and mean height difference
}
\description{
Return the height difference of the input dataframe withe the mean height of the dataframe.
}
\author{
Tobias Riebe, Michael Graber
}
