% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sports_plot.R
\name{create_sports_plot}
\alias{create_sports_plot}
\title{create_sports_plot}
\usage{
create_sports_plot(data)
}
\arguments{
\item{data}{Give a dataframe with team names with wins and losses}
}
\value{
A ggplot
}
\description{
create_sports_plot
}
\examples{
create_sports_plot(data)

}
