% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Sports Teams Performance Data}
\format{
## `data`
A data frame with 4 rows and 4 columns:
\describe{
  \item{X}{An integer indicating the row number.}
  \item{team}{The name of the sports team.}
  \item{wins}{The number of games the team has won.}
  \item{losses}{The number of games the team has lost.}
}
}
\source{
Simulated data
}
\usage{
data
}
\description{
This dataset contains information about the performance of various sports teams, including their number of wins and losses.
}
\keyword{datasets}
