% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean}
\alias{mean}
\title{Calculates the arithmetic mean}
\usage{
mean(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
an integer
}
\description{
Calculates the arithmetic mean and round on the 3rd digit
}
\author{
Michael Graber
}
