% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myCalc.R
\name{my_calc}
\alias{my_calc}
\title{Title}
\usage{
my_calc(x, f)
}
\arguments{
\item{x}{A numeric vector for calculation}

\item{f}{A function to be applied to x}
}
\value{
Result of calculation
}
\description{
Title
}
\examples{
my_calc(1:10, mean)
}
