% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_function.R
\name{my_function}
\alias{my_function}
\title{My example function}
\usage{
my_function(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The sum of x
}
\description{
This function does something interesting
}
\examples{
my_function(1:10)
}
