% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permtest.r
\name{permtest}
\alias{permtest}
\title{Perform a permutation test on a linreg object}
\usage{
permtest(model, nperm = 99)
}
\arguments{
\item{model}{A linreg model}

\item{nperm}{Number of permutations}
}
\value{
Silently returns a list that includes the
following elements:
\describe{
  \item{perms}{A matrix containing the values for all coefficients
  on each permutation}
  \item{coefficients}{A matrix of coefficients and their permutation
  based p-values}}
}
\description{
Permutes the observations and calculates a p-value
for every coefficient based on those permutations.
}
\details{
Takes a linreg object, permutes the observations
and uses the estimated null distributions to obtain
a two-sided p-value of each coefficient in the
linreg model.

It ignores cases of undefined values in permutations.
}
\seealso{
\link{linreg}
}
\author{
Sur Herrera Paredes
}
