% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.r
\name{print.linreg}
\alias{print.linreg}
\title{Print linreg object}
\usage{
\method{print}{linreg}(x, digits = 2)
}
\arguments{
\item{x}{Linreg object}

\item{digits}{Number of digits to print}
}
\value{
Invisibly a matrix of coefficient estimates and
standard errors
}
\description{
Pretty display of linreg object
}
\details{
This is an S3 method for the generic \link{print} function.

It builds a matrix of coefficients and standard
error and it prints it
}
\seealso{
\link{print} \link{linreg}
}
\author{
Sur Herrera Paredes
}
