% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{students}
\alias{students}
\title{Dataframe withe some attributes of students.}
\format{A data frame with five variables:
\describe{
  \item{age}{age, in years}
  \item{weight}{weight, in kg}
  \item{height}{height, in m}
  \item{sex}{sex, coded as factors}
  \item{name}{names}
 }}
\usage{
students
}
\description{
Dataframe withe some attributes of students.
}
\keyword{datasets}
