% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.r
\name{summary.linreg}
\alias{summary.linreg}
\title{Summarize linreg model fit}
\usage{
\method{summary}{linreg}(object)
}
\arguments{
\item{object}{A linreg object}
}
\value{
Invisibly return a matrix of coefficient and
test results
}
\description{
Provides a full summary of the lireg
fit and performs t-tests on the coefficients
}
\details{
This function is a n S3 method of the generic \link{summary}
function.

Takes a linreg object, calculates the degrees of freedom
and performs a two-tailed t-test on each coefficient where
the null hypohtesis is that its real value is zero.
}
\seealso{
\link{summary} \link{linreg}
}
\author{
Sur Herrera Paredes
}
